TriggerTable created
